move to legacy: stream_data_interfaces
