+++ weave/legacy/run_streamtable_span.py
@@ -1,7 +1,7 @@
 import typing
 from typing import Iterable
 
-from weave import stream_data_interfaces
+from weave.legacy import stream_data_interfaces
 from weave.eager import WeaveIter
 from weave.legacy import artifact_wandb, uris
 from weave.legacy.run import Run
