+++ weave/legacy/stream_data_interfaces.py
